requirements.txt
annotated-types==0.7.0
audioop-lts==0.2.2
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
diskcache==5.6.3
dlib==20.0.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numba==0.61.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.0
pooch==1.8.2
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
typing-inspection==0.4.1
typing_extensions==4.15.0
ultralytics==8.3.195
ultralytics-thop==2.0.17
urllib3==2.5.0
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2